requirements.txt
opencv-python==4.8.0.74
numpy==1.24.3
mss==9.0.1
pyautogui==0.9.54
websocket-client==1.6.1